cleanups and refactoring
